requirements.txt
opencv-python
easyocr
transformers
torch
numpy
Pillow
